+++ src/ska_dlm_client/utils.py
@@ -10,7 +10,6 @@ import logging
 import os
 from pathlib import Path
 
-logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)