+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Updated
 
 * Running tests with server image 2.0.0.
+* The ConfigDB watcher now watches DataProductPersist Flows and checks the status of the related DataProduct Flow.
 
 ## 2.0.0